Update the parser logic of the autonomous system of TAXII server for the
Resolution of the parser logic which required the "number" key in format of the Number instead of the string. As per the standard TAXII feed provides the AS number in type of number rather than string. To handle the logic we have update the parser logic.

Official Documentation : https://docs.oasis-open.org/cti/stix/v2.1/os/stix-v2.1-os.pdf

Page Number : 133
Object : 6.2 Autonomous System (AS) Object